pyproject.toml
[tool.poetry]
name = "yaml-extras"
version = "0.1.0"
description = "A collection of YAML utilities"
authors = ["David Sillman <dsillman2000@gmail.com>"]
packages = [{ include = "yaml_extras" }]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
pyyaml-include = "^2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
